Allow additional module inclusion [CBLUR-66]
Additional modules can be included at configure time by setting
the following CMake variables:
    COREBLURON_ADDITIONAL_MECHS
	Path to a file containing a list of mod files to include.
        One .mod file per line, which can be specified with an
        absolute path name, or as a bare file name.
    COREBLURON_ADDITIONAL_MECHPATH
        List of directories to search for mod files specified in
        COREBLURON_ADDITIONAL_MECHS

Mod files specified in COREBLURON_ADDITIONAL_MECHS or in the
coreneuron_modlist.txt file that cannot be located are reported
with a warning rather than with a fatal error.

In addition, the mandatory modules are compiled in regardless
of the setting of COREBLURON_NEURODAMUS_MECHS.

Modifying the coreneuron_modlist.txt file or the file specified
in COREBLURON_ADDITIONAL_MECHS correctly triggers a reconfigure.

Additions:
1. Add unit test for the functionality based on
   CMake reconfiguration.
2. Move BSD License text from generated mod_func.c to
   the generating perl script mod_func.c.pl
3. Temporarily switch .gitexternals for CMake to point
   to a branch on Sam's fork of Eyescale/CMake pending
   PR inclusion.

Change-Id: Id3f6f8c9969f7e49f40188691d9c7260de116e74