Support for NRN_SOA_PAD > 1
Requires corresponding changes in mod2c_core.

NRN_SOA_PAD > 1 is relevant only for LAYOUT 0 .
The primary computation is to modify the integer values of the
Memb_list.pdata array when those integers are indices
into NrnThread._data . The indices handled can point
into nt._actual_area, nt._actual_v, and various ion data.
The MATRIX_LAYOUT definition can be used to turn off padding for the
nt._actual_* arrays.
To help with debugging with gdb, a #if 0 ... #endif fragment was
added to mina1.cpp which, when switched on, allows floating errors to
become exceptions.

Change-Id: I1b678ab17e809c33ec58c90507f45cea24e9f365